fix: Prevent bluetooth from preventing laptops from staying suspended
I've been hearing that the bluetooth service causes a number of laptops, including the oryp5, to be unable to remain suspended. A script similar to this is being passed around as a workaround to the problem, so it might be useful to include this workaround by default. This will disable and stop the bluetooth service immediately pre-suspension, and then re-enable and start it post-suspension. A file is used to determine if the bluetooth service was enabled prior to suspension, so that it will only be enabled and started if it was previously-enabled