fix(proprietary): add the Deb822-format repository on groovy
Since the ISO is being built with the DEB822-format sources (including the Apps repo),
the postinst script needs to be adapted to add it as well if it isn't found. Otherwise,
the ISO ends up getting build with two copies of the repo, causing APT errors in the
ISO and installed system.