chore: Increase max uncompressed size and portion
It seems that because we're using zstd, and zstd has a much higher
compression ratio than lzo, it would make sense to increase the default
to 16 GB of uncompressed memory. 8 GB limited the size of the zram block
device to no greater than ~2.4 GB after compression.

Fedora has recently changed their zram configuration to 100% of RAM
size, and I've been doing well with this configuration for a few days of
compiling and web browsing on a system with 16 GB.